% Generated by roxygen2 (4.0.1): do not edit by hand
\name{asISO8601}
\alias{asISO8601}
\title{Returns date time string in ISO8601 format}
\usage{
asISO8601(dt)
}
\arguments{
\item{dt}{Date time object which can be converted to a POSIXlt object.}
}
\value{
Character string in ISO8601 format.
}
\description{
Converts an object which can be converted to a POSIXlt object to a ISO8601 date time string.
}
\examples{
asISO8601("2014-01-01")
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}

