% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFSRequest.R
\name{GMLFile}
\alias{GMLFile}
\title{A class for providing a file name reference to a GML document}
\description{
A class for providing a file name reference to a GML document.
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}
\section{Super class}{
\code{\link[rwfs:WFSRequest]{rwfs::WFSRequest}} -> \code{GMLFile}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GMLFile$new()}}
\item \href{#method-getDataSource}{\code{GMLFile$getDataSource()}}
\item \href{#method-clone}{\code{GMLFile$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="rwfs" data-topic="WFSRequest" data-id="print">}\href{../../rwfs/html/WFSRequest.html#method-print}{\code{rwfs::WFSRequest$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMLFile$new(srcFile)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDataSource"></a>}}
\if{latex}{\out{\hypertarget{method-getDataSource}{}}}
\subsection{Method \code{getDataSource()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMLFile$getDataSource()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMLFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
