% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFSRequest.R
\name{WFSCachingRequest}
\alias{WFSCachingRequest}
\title{An abstract class for building a URL reference to a WFS with a caching}
\description{
The abstract method \code{getURL} must be overloaded in a subclass to provide a request URL to a WFS service.
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}
\section{Super classes}{
\code{\link[rwfs:WFSRequest]{rwfs::WFSRequest}} -> \code{\link[rwfs:WFSStreamingRequest]{rwfs::WFSStreamingRequest}} -> \code{WFSCachingRequest}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-getDataSource}{\code{WFSCachingRequest$getDataSource()}}
\item \href{#method-clone}{\code{WFSCachingRequest$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="rwfs" data-topic="WFSRequest" data-id="print">}\href{../../rwfs/html/WFSRequest.html#method-print}{\code{rwfs::WFSRequest$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rwfs" data-topic="WFSStreamingRequest" data-id="getCapabilities">}\href{../../rwfs/html/WFSStreamingRequest.html#method-getCapabilities}{\code{rwfs::WFSStreamingRequest$getCapabilities()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rwfs" data-topic="WFSStreamingRequest" data-id="getFeature">}\href{../../rwfs/html/WFSStreamingRequest.html#method-getFeature}{\code{rwfs::WFSStreamingRequest$getFeature()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rwfs" data-topic="WFSStreamingRequest" data-id="getParameters">}\href{../../rwfs/html/WFSStreamingRequest.html#method-getParameters}{\code{rwfs::WFSStreamingRequest$getParameters()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rwfs" data-topic="WFSStreamingRequest" data-id="setParameters">}\href{../../rwfs/html/WFSStreamingRequest.html#method-setParameters}{\code{rwfs::WFSStreamingRequest$setParameters()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rwfs" data-topic="WFSStreamingRequest" data-id="setPath">}\href{../../rwfs/html/WFSStreamingRequest.html#method-setPath}{\code{rwfs::WFSStreamingRequest$setPath()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDataSource"></a>}}
\if{latex}{\out{\hypertarget{method-getDataSource}{}}}
\subsection{Method \code{getDataSource()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WFSCachingRequest$getDataSource()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WFSCachingRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
