% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{WFSClient-class}
\alias{WFSClient}
\alias{WFSClient-class}
\alias{getRasterURL}
\alias{importRaster}
\title{WFS client abstract reference class}
\description{
WFS client abstract reference class
}
\section{Methods}{

\describe{
\item{\code{getLayer(request, layer, crs = NULL, swapAxisOrder = FALSE, parameters)}}{Returns layer data from a WFS query.}

\item{\code{getRaster(request, parameters)}}{Returns raster from WFS query.}

\item{\code{getRasterURL(request, parameters)}}{Returns raster URL from a WFS query response.}

\item{\code{importRaster(destFile)}}{Imports raster from downloaded file.}

\item{\code{listLayers(request)}}{Lists data layers from a WFS query.}
}}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}

