% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFSClient.R
\docType{class}
\name{WFSClient}
\alias{WFSClient}
\title{Class represeting a WFS client}
\format{\code{\link{R6Class}} object.}
\description{
An abstract class to represent OGC's WFS client in R. Other client classes
in this package inherit this this class.
}
\section{Fields}{

\describe{
\item{\code{test}}{}
}}

\section{Methods}{

\describe{
  \item{\code{new(request)}}{This method is used to create object of this 
        class with \code{request} as the request object containing WFS
        connection information and methods. NOTE: as this is abstract class,
        you shouldn't be creating instances of it.}
  \item{setRequest(request)}{Set client's request object to \code{request}, 
        which must inherit from \code{\link{WFSRequest}}.}
  \item{listLayers()}{Not implemented in this abstract class, but it classes
        inheriting this class.}
  \item{getLayer}{Not implemented in this abstract class, but it classes
        inheriting this class.}
  \item{getRaster}{Get a raster layer from WFS }
}
}

\seealso{
\code{\link{WFSStreamingClient}}, \code{\link{WFSCachingClient}}, 
         \code{\link{WFSRequest}}
}
\author{
Jussi Jousimo \email{jvj@iki.fi},
        Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}
\keyword{datasets}
