% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{WFSFileClient}
\alias{WFSFileClient}
\title{Reads response from a file instead of a WFS service directly}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x454a070> 
 - attr(*, "name")= chr "WFSFileClient_generator"
}}
\usage{
WFSFileClient
}
\description{
Dispatches WFS request, saves response to a file and parses the file.
The absract method \code{\link{getRasterURL}} should be overloaded for raster queries and possibly
\code{\link{importRaster}} as well.
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}
\seealso{
\code{\link{WFSRequest}}, \code{\link{WFSStreamClient}}
}
\keyword{datasets}

