% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{WFSRequest-class}
\alias{WFSRequest}
\alias{WFSRequest-class}
\alias{getURL}
\title{An abstract class to build WFS request URL's}
\description{
This class should be inherited and the abstract method \code{getURL} overloaded to provide WFS request URL's.
}
\section{Methods}{

\describe{
\item{\code{getParametersString()}}{Returns WFS request URL parameters as a string.}

\item{\code{getPathString()}}{Returns WFS request URL path as a string.}

\item{\code{getURL()}}{Returns WFS request URL.}

\item{\code{setParameters(...)}}{Sets WFS request URL parameters.}

\item{\code{setPath(path)}}{Sets WFS request URL path.}
}}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}
\seealso{
\code{\link{WFSStreamClient}}, \code{\link{WFSFileClient}}
}

