% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{WFSRequest}
\alias{WFSRequest}
\alias{getURL}
\title{An abstract class to build WFS request URL's}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x44dc688> 
 - attr(*, "name")= chr "WFSRequest_generator"
}}
\usage{
WFSRequest
}
\description{
This class should be inherited and the abstract method \code{getURL} overloaded to provide WFS request URL's.
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}
\seealso{
\code{\link{WFSStreamClient}}, \code{\link{WFSFileClient}}
}
\keyword{datasets}

