% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{WFSStreamClient}
\alias{WFSStreamClient}
\title{Streams response from a WFS}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x4536550> 
 - attr(*, "name")= chr "WFSStreamClient_generator"
}}
\usage{
WFSStreamClient
}
\description{
Dispatches WFS request and parses response from the stream. Provides a caching mechanism
for the same subsequent queries. The absract method \code{\link{getRasterURL}} should be overloaded for
raster queries and possibly \code{\link{importRaster}} as well.
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}
\seealso{
\code{\link{WFSRequest}}, \code{\link{WFSFileClient}}
}
\keyword{datasets}

