% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFSClient.R
\name{WFSStreamingClient}
\alias{WFSStreamingClient}
\title{Streams response from a WFS}
\description{
Dispatches a WFS request and parses response from the stream directly.
}
\seealso{
\code{\link{WFSRequest}}, \code{\link{WFSCachingClient}}
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}
\section{Super class}{
\code{\link[rwfs:WFSClient]{rwfs::WFSClient}} -> \code{WFSStreamClient}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-listLayers}{\code{WFSStreamingClient$listLayers()}}
\item \href{#method-getLayer}{\code{WFSStreamingClient$getLayer()}}
\item \href{#method-clone}{\code{WFSStreamingClient$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="rwfs" data-topic="WFSClient" data-id="getRaster">}\href{../../rwfs/html/WFSClient.html#method-getRaster}{\code{rwfs::WFSClient$getRaster()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rwfs" data-topic="WFSClient" data-id="initialize">}\href{../../rwfs/html/WFSClient.html#method-initialize}{\code{rwfs::WFSClient$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rwfs" data-topic="WFSClient" data-id="setRequest">}\href{../../rwfs/html/WFSClient.html#method-setRequest}{\code{rwfs::WFSClient$setRequest()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-listLayers"></a>}}
\if{latex}{\out{\hypertarget{method-listLayers}{}}}
\subsection{Method \code{listLayers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WFSStreamingClient$listLayers()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getLayer"></a>}}
\if{latex}{\out{\hypertarget{method-getLayer}{}}}
\subsection{Method \code{getLayer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WFSStreamingClient$getLayer(layer, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WFSStreamingClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
