% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFSRequest.R
\name{WFSStreamingRequest}
\alias{WFSStreamingRequest}
\title{An abstract class for building a URL reference to a WFS}
\description{
An abstract class for building a URL reference to a WFS.
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}
\section{Super class}{
\code{\link[rwfs:WFSRequest]{rwfs::WFSRequest}} -> \code{WFSStreamingRequest}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-getParameters}{\code{WFSStreamingRequest$getParameters()}}
\item \href{#method-setPath}{\code{WFSStreamingRequest$setPath()}}
\item \href{#method-setParameters}{\code{WFSStreamingRequest$setParameters()}}
\item \href{#method-getCapabilities}{\code{WFSStreamingRequest$getCapabilities()}}
\item \href{#method-getFeature}{\code{WFSStreamingRequest$getFeature()}}
\item \href{#method-clone}{\code{WFSStreamingRequest$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="rwfs" data-topic="WFSRequest" data-id="getDataSource">}\href{../../rwfs/html/WFSRequest.html#method-getDataSource}{\code{rwfs::WFSRequest$getDataSource()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rwfs" data-topic="WFSRequest" data-id="print">}\href{../../rwfs/html/WFSRequest.html#method-print}{\code{rwfs::WFSRequest$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getParameters"></a>}}
\if{latex}{\out{\hypertarget{method-getParameters}{}}}
\subsection{Method \code{getParameters()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WFSStreamingRequest$getParameters()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setPath"></a>}}
\if{latex}{\out{\hypertarget{method-setPath}{}}}
\subsection{Method \code{setPath()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WFSStreamingRequest$setPath(path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setParameters"></a>}}
\if{latex}{\out{\hypertarget{method-setParameters}{}}}
\subsection{Method \code{setParameters()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WFSStreamingRequest$setParameters(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCapabilities"></a>}}
\if{latex}{\out{\hypertarget{method-getCapabilities}{}}}
\subsection{Method \code{getCapabilities()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WFSStreamingRequest$getCapabilities(version = "1.0.0", ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getFeature"></a>}}
\if{latex}{\out{\hypertarget{method-getFeature}{}}}
\subsection{Method \code{getFeature()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WFSStreamingRequest$getFeature(version = "1.0.0", typeNames, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WFSStreamingRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
