% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{GMLFile}
\alias{GMLFile}
\title{A class for providing a file name reference to a GML document}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x4502b20> 
 - attr(*, "name")= chr "GMLFile_generator"
}}
\usage{
GMLFile
}
\description{
A class for providing a file name reference to a GML document
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}
\keyword{datasets}

