% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{rwfs}
\alias{rwfs}
\alias{rwfs-package}
\title{WFS client for R}
\description{
This R package provides a client to access Web Feature Services (WFS) (\url{http://www.opengeospatial.org/standards/wfs}).
}
\details{
The client relies on the GDAL (\url{http://www.gdal.org/}) library and the \pkg{rgdal}
(\url{https://cran.r-project.org/package=rgdal}) package, which should be installed first.
 
See examples in the \pkg{fmi} (\url{http://www.github.com/rOpenGov/fmi}) and the \pkg{gisfin} packages
(\url{http://www.github.com/rOpenGov/gisfin}) how to inherit the abstract R6 classes
(\url{https://cran.r-project.org/package=R6}) provided by the package.
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}
