% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{WFSCachingRequest}
\alias{WFSCachingRequest}
\title{An abstract class for building a URL reference to a WFS with a caching}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x44566b0> 
 - attr(*, "name")= chr "WFSCachingRequest_generator"
}}
\usage{
WFSCachingRequest
}
\description{
The abstract method \code{getURL} must be overloaded in a subclass to provide a request URL to a WFS service.
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}
\keyword{datasets}

