% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFSClient.R
\docType{data}
\name{WFSClient}
\alias{WFSClient}
\title{WFS client abstract class}
\description{
An abstract class to represent OGC's WFS.
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}
\seealso{
\code{\link{WFSStreamingClient}}, \code{\link{WFSCachingClient}}, \code{\link{WFSRequest}}
}
\keyword{datasets}

