% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{WFSClient}
\alias{WFSClient}
\title{WFS client abstract class}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x4453918> 
 - attr(*, "name")= chr "WFSClient_generator"
}}
\usage{
WFSClient
}
\description{
WFS client abstract class
}
\section{Methods}{

\itemize{
 \item \code{getDailyWeather}: Returns daily weather time-seris
 \item \code{getMonthlyWeatherRaster}: Returns daily weather raster
}
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}
\seealso{
\code{\link{WFSStreamingClient}}, \code{\link{WFSCachingClient}}, \code{\link{WFSRequest}}
}
\keyword{datasets}

