% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{WFSClient}
\alias{WFSClient}
\alias{getRasterURL}
\alias{importRaster}
\title{WFS client abstract reference class}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x4522f88> 
 - attr(*, "name")= chr "WFSClient_generator"
}}
\usage{
WFSClient
}
\description{
WFS client abstract reference class
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}
\keyword{datasets}

