% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFSClient.R
\name{WFSClient}
\alias{WFSClient}
\title{Class represeting a WFS client}
\format{
\code{\link{R6Class}} object.
}
\description{
An abstract class to represent OGC's WFS client in R. Other client classes
in this package inherit this this class.
}
\section{Methods}{

\describe{
  \item{\code{new(request)}}{This method is used to create object of this 
        class with \code{request} as the request object containing WFS
        connection information and methods. NOTE: as this is abstract class,
        you shouldn't be creating instances of it.}
  \item{setRequest(request)}{Set client's request object to \code{request}, 
        which must inherit from \code{\link{WFSRequest}}.}
  \item{listLayers()}{Not implemented in this abstract class, but it classes
        inheriting this class.}
  \item{getLayer}{Not implemented in this abstract class, but it classes
        inheriting this class.}
  \item{getRaster}{Get a raster layer from WFS }
}
}

\seealso{
\code{\link{WFSStreamingClient}}, \code{\link{WFSCachingClient}}, 
         \code{\link{WFSRequest}}
}
\author{
Jussi Jousimo \email{jvj@iki.fi},
        Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{test}}{}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{test}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{WFSClient$new()}}
\item \href{#method-setRequest}{\code{WFSClient$setRequest()}}
\item \href{#method-listLayers}{\code{WFSClient$listLayers()}}
\item \href{#method-getLayer}{\code{WFSClient$getLayer()}}
\item \href{#method-getRaster}{\code{WFSClient$getRaster()}}
\item \href{#method-clone}{\code{WFSClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WFSClient$new(request)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setRequest"></a>}}
\if{latex}{\out{\hypertarget{method-setRequest}{}}}
\subsection{Method \code{setRequest()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WFSClient$setRequest(request)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-listLayers"></a>}}
\if{latex}{\out{\hypertarget{method-listLayers}{}}}
\subsection{Method \code{listLayers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WFSClient$listLayers()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getLayer"></a>}}
\if{latex}{\out{\hypertarget{method-getLayer}{}}}
\subsection{Method \code{getLayer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WFSClient$getLayer(layer, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getRaster"></a>}}
\if{latex}{\out{\hypertarget{method-getRaster}{}}}
\subsection{Method \code{getRaster()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WFSClient$getRaster(parameters)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WFSClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
