% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{WFSFileClient-class}
\alias{WFSFileClient}
\alias{WFSFileClient-class}
\title{Reads response from a file instead of a WFS service directly.}
\description{
Reads response from a file instead of a WFS service directly.
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}

