% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFSRequest.R
\name{WFSRequest}
\alias{WFSRequest}
\title{An abstract class for referencing a WFS or a GML document}
\description{
This class should be inherited and the abstract method \code{getDataSource} overloaded
in a subclass to provide a reference.
}
\seealso{
\code{\link{WFSClient}}, \code{\link{GMLFile}}
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-getDataSource}{\code{WFSRequest$getDataSource()}}
\item \href{#method-print}{\code{WFSRequest$print()}}
\item \href{#method-clone}{\code{WFSRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDataSource"></a>}}
\if{latex}{\out{\hypertarget{method-getDataSource}{}}}
\subsection{Method \code{getDataSource()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WFSRequest$getDataSource()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WFSRequest$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WFSRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
