% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{WFSRequest}
\alias{WFSRequest}
\title{An abstract class for referencing a WFS or a GML document}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x44c84e0> 
 - attr(*, "name")= chr "WFSRequest_generator"
}}
\usage{
WFSRequest
}
\description{
This class should be inherited and the abstract method \code{getDataSource} overloaded
in a subclass to provide a reference.
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}
\seealso{
\code{\link{WFSClient}}, \code{\link{GMLFile}}
}
\keyword{datasets}

