% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{WFSStreamingClient}
\alias{WFSStreamingClient}
\title{Streams response from a WFS}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x43e9c48> 
 - attr(*, "name")= chr "WFSStreamClient_generator"
}}
\usage{
WFSStreamingClient
}
\description{
Dispatches a WFS request and parses response from the stream directly.
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}
\seealso{
\code{\link{WFSRequest}}, \code{\link{WFSCachingClient}}
}
\keyword{datasets}

