% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{WFSStreamingRequest}
\alias{WFSStreamingRequest}
\title{An abstract class for building a URL reference to a WFS}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x44dc270> 
 - attr(*, "name")= chr "WFSStreamingRequest_generator"
}}
\usage{
WFSStreamingRequest
}
\description{
An abstract class for building a URL reference to a WFS
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}
\keyword{datasets}

